+++ certbot/certbot/util.py
@@ -17,6 +17,7 @@ import sys
 import configargparse
 import six
 
+from acme.magic_typing import Text
 from acme.magic_typing import Tuple
 from acme.magic_typing import Union
 from certbot import errors
@@ -401,14 +402,14 @@ def get_python_os_info(pretty=False):
         try:
             proc = subprocess.Popen(
                 ["/usr/bin/sw_vers", "-productVersion"],
-                stdout=subprocess.PIPE,
+                stdout=subprocess.PIPE, stderr=subprocess.PIPE,
                 universal_newlines=True,
                 env=env_no_snap_for_external_calls(),
             )
         except OSError:
             proc = subprocess.Popen(
                 ["sw_vers", "-productVersion"],
-                stdout=subprocess.PIPE,
+                stdout=subprocess.PIPE, stderr=subprocess.PIPE,
                 universal_newlines=True,
                 env=env_no_snap_for_external_calls(),
             )
@@ -585,11 +586,9 @@ def is_wildcard_domain(domain):
     :rtype: bool
 
     """
+    wildcard_marker = b"*."  # type: Union[Text, bytes]
     if isinstance(domain, six.text_type):
         wildcard_marker = u"*."
-    else:
-        wildcard_marker = b"*."
-
     return domain.startswith(wildcard_marker)