+++ tools/snap/build_remote.py
@@ -6,7 +6,8 @@ from multiprocessing import Pool, Process, Manager, Event
 import re
 import subprocess
 import sys
-from os.path import join, realpath, dirname, basename
+import tempfile
+from os.path import join, realpath, dirname, basename, exists
 
 
 CERTBOT_DIR = dirname(dirname(dirname(realpath(__file__))))
@@ -16,37 +17,53 @@ PLUGINS = [basename(path) for path in glob.glob(join(CERTBOT_DIR, 'certbot-dns-*
 def _execute_build(target, archs, status, workspace):
     process = subprocess.Popen([
         'snapcraft', 'remote-build', '--launchpad-accept-public-upload', '--recover', '--build-on', ','.join(archs)
-    ], stdout=subprocess.PIPE, stderr=subprocess.PIPE, universal_newlines=True, cwd=workspace)
+    ], stdout=subprocess.PIPE, stderr=subprocess.STDOUT, universal_newlines=True, cwd=workspace)
 
+    process_output = []
     for line in process.stdout:
+        process_output.append(line)
         _extract_state(target, line, status)
 
-    return process.wait()
+    return process.wait(), process_output
 
 
-def _build_snap(target, archs, status):
+def _build_snap(target, archs, status, lock):
     status[target] = {arch: '...' for arch in archs}
 
     if target == 'certbot':
         workspace = CERTBOT_DIR
     else:
         workspace = join(CERTBOT_DIR, target)
-        subprocess.check_output(
-            ('"{0}" tools/strip_hashes.py letsencrypt-auto-source/pieces/dependency-requirements.txt '
-             '| grep -v python-augeas > "{1}/snap-constraints.txt"').format(sys.executable, workspace),
-            shell=True, cwd=CERTBOT_DIR)
 
     retry = 3
     while retry:
-        exit_code = _execute_build(target, archs, status, workspace)
+        exit_code, process_output = _execute_build(target, archs, status, workspace)
 
         print(f'Build {target} for {",".join(archs)} (attempt {4-retry}/3) ended with exit code {exit_code}.')
         sys.stdout.flush()
 
-        # Retry if the snapcraft remote-build command has been interrupted.
-        if exit_code == 0 and 'Failed to build' not in status[target].values():
-            break
+        with lock:
+            dump_output = exit_code != 0
+            failed_archs = [arch for arch in archs if status[target][arch] == 'Failed to build']
+            if exit_code == 0 and not failed_archs:
+                # We expect to have all target snaps available, or something bad happened.
+                snaps_list = glob.glob(join(workspace, '*.snap'))
+                if not len(snaps_list) == len(archs):
+                    print(f'Some of the expected snaps for a successful build are missing (current list: {snaps_list}).')
+                    dump_output = True
+                else:
+                    break
+            if failed_archs:
+                # We expect each failed build to have a log file, or something bad happened.
+                for arch in failed_archs:
+                    if not exists(join(workspace, f'{target}_{arch}.txt')):
+                        dump_output = True
+                        print(f'Missing output on a failed build {target} for {arch}.')
+            if dump_output:
+                print(f'Dumping snapcraft remote-build output build for {target}:')
+                print('\n'.join(process_output))
 
+        # Retry the remote build if it has been interrupted (non zero status code) or if some builds have failed.
         retry = retry - 1
 
     return {target: workspace}
@@ -99,8 +116,12 @@ def _dump_results(targets, archs, status, workspaces):
             if result != 'Successfully built':
                 failures = True
 
-                with open(join(workspaces[target], '{0}_{1}.txt'.format(target, arch))) as file_h:
-                    build_output = file_h.read()
+                build_output_path = join(workspaces[target], f'{target}_{arch}.txt')
+                if not exists(build_output_path):
+                    build_output = f'No output has been dumped by snapcraft remote-build.'
+                else:
+                    with open(join(workspaces[target], '{0}_{1}.txt'.format(target, arch))) as file_h:
+                        build_output = file_h.read()
 
                 print('Output for failed build target={0} arch={1}'.format(target, arch))
                 print('-------------------------------------------')
@@ -120,7 +141,7 @@ def main():
     parser = argparse.ArgumentParser()
     parser.add_argument('targets', nargs='+', choices=['ALL', 'DNS_PLUGINS', 'certbot', *PLUGINS],
                         help='the list of snaps to build')
-    parser.add_argument('--archs', nargs='+', choices=['amd64', 'arm64', 'armhf'], default='amd64',
+    parser.add_argument('--archs', nargs='+', choices=['amd64', 'arm64', 'armhf'], default=['amd64'],
                         help='the architectures for which snaps are built')
     args = parser.parse_args()
 
@@ -135,6 +156,12 @@ def main():
         targets.remove('DNS_PLUGINS')
         targets.update(PLUGINS)
 
+    # If we're building anything other than just Certbot, we need to
+    # generate the snapcraft files for the DNS plugins.
+    if targets != set(('certbot',)):
+        subprocess.run(['tools/snap/generate_dnsplugins_all.sh'],
+                       check=True, cwd=CERTBOT_DIR)
+
     print('Start remote snap builds...')
     print(f' - archs: {", ".join(archs)}')
     print(f' - projects: {", ".join(sorted(targets))}')
@@ -142,12 +169,13 @@ def main():
 
     with Manager() as manager, Pool(processes=len(targets)) as pool:
         status = manager.dict()
+        lock = manager.Lock()
 
         stop_event = Event()
         state_process = Process(target=_dump_status, args=(archs, status, stop_event))
         state_process.start()
 
-        async_results = [pool.apply_async(_build_snap, (target, archs, status)) for target in targets]
+        async_results = [pool.apply_async(_build_snap, (target, archs, status, lock)) for target in targets]
 
         workspaces = {}
         for async_result in async_results: